+++ lab8/joystick.h
@@ -4,6 +4,6 @@
 extern int PrevStat;
 extern int state;
 void joystick_init();
-uint16_t joystick_adc(uint8_t dim);
+double joystick_adc(uint8_t dim);
 
 #endif
